Final Migration to Assembly C#
