+++ SharedProject/SharedSource/Configuration.cs
